+++ compiler/src/instructions/AddressResolver.ts
@@ -1,20 +1,20 @@
 import { InstructionBase } from ".";
-import { IBindableValue, IScope } from "../types";
+import { IScope, TLineRef } from "../types";
 
 export class AddressResolver extends InstructionBase {
   public get hidden() {
     return true;
   }
   public set hidden(value) {}
-  bonds: IBindableValue<number | null>[];
-  constructor(...bonds: IBindableValue<number | null>[]) {
+  bonds: TLineRef[];
+  constructor(...bonds: TLineRef[]) {
     super();
     this.bonds = bonds;
   }
   resolve(i: number) {
     for (const literal of this.bonds) literal.data = i;
   }
-  bind(bond: IBindableValue<number | null>) {
+  bind(bond: TLineRef) {
     this.bonds.push(bond);
   }
   bindBreak(scope: IScope) {