+++ compiler/src/instructions/ControlInstructions.ts
@@ -1,10 +1,10 @@
-import { EInstIntent, IBindableValue } from "../types";
+import { EInstIntent, TLineRef } from "../types";
 import { EJumpKind, JumpInstruction } from "./JumpInstruction";
 
 export class BreakInstruction extends JumpInstruction {
   intent = EInstIntent.break;
 
-  constructor(address: IBindableValue<number | null>) {
+  constructor(address: TLineRef) {
     super(address, EJumpKind.Always);
   }
 }
@@ -12,7 +12,7 @@ export class BreakInstruction extends JumpInstruction {
 export class ContinueInstruction extends JumpInstruction {
   intent = EInstIntent.continue;
 
-  constructor(address: IBindableValue<number | null>) {
+  constructor(address: TLineRef) {
     super(address, EJumpKind.Always);
   }
 }